requirements.txt
Flask==2.3.3
PyJWT==2.8.0
python-dotenv==1.0.0
requests==2.31.0

